tests: Fix all
